ci: fix clippy::needless_return + 6 OAuth-module lints to unblock pi-rollout
Lint Rust step on rust-1.93/1.95 toolchains in CI rejected the OAuth
modules (added by F16pi-anthropic-oauth + F16pi-openai-codex). Local
clippy passed because the workspace was pinned to an earlier toolchain.

- anthropic.rs: drop `as u64` cast on already-u64 (`unnecessary_cast`).
- openai_codex.rs: `% 4 != 0` → `!.is_multiple_of(4)` (`manual_is_multiple_of`).
- anthropic_tests.rs (×3): struct-update form instead of
  `let mut .. = ::default(); field = …` (`field_reassign_with_default`).
- anthropic_tests.rs: collapse `if let … if let …` into `&&` chain (`collapsible_if`).
- anthropic_tests.rs: drop `body.as_bytes().len()` (`needless_as_bytes`).
- squeezy-cli/auth.rs: drop `return Ok(());` in `open_browser` cfg blocks (`needless_return`).

+++ crates/squeezy-llm/src/oauth/anthropic.rs
@@ -115,7 +115,7 @@ impl PersistedTokens {
     pub fn from_token_response(response: &TokenResponse, now_ms: u64) -> Self {
         let expires_at_unix_ms = now_ms
             .saturating_add(response.expires_in.saturating_mul(1000))
-            .saturating_sub((REFRESH_LEAD_TIME.as_secs() as u64).saturating_mul(1000));
+            .saturating_sub(REFRESH_LEAD_TIME.as_secs().saturating_mul(1000));
         Self {
             access_token: response.access_token.clone(),
             refresh_token: response.refresh_token.clone(),
